requirements.txt
python-telegram-bot[job-queue]==22.7
yfinance==1.3.0